refatorar(testes): substituir FakePlatformViewsController por FakeGoogleMapPlatformViewsController e mover para um novo arquivo para melhorar a estrutura do teste
